apkBot/requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
comtypes==1.1.11
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
filelock==3.7.1
Flask==2.1.2
gunicorn==20.1.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.0
MarkupSafe==2.1.1
outcome==1.2.0
parsel==1.6.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
PySocks==1.7.1
python-telegram-bot==13.12
pyttsx3==2.90
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
queuelib==1.6.2
requests==2.28.0
requests-file==1.5.1
Scrapy==2.6.1
selenium==4.2.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tldextract==3.3.0
tornado==6.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
w3lib==1.22.0
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.8.0
zope.interface==5.4.0
